+++ src/display/components/Spinner.js
@@ -1,0 +1,8 @@
+import React from "react";
+import loader from "../styles/SpinnerStyles.css";
+
+const Spinner = () => {
+  return <div class="loader">Loading...</div>;
+};
+
+export default Spinner;
