+++ src/features/add/Search.js
@@ -1,0 +1,26 @@
+import React, { useState } from "react";
+
+const Search = ({ getQuery }) => {
+  const [text, setText] = useState("");
+
+  const onChange = (q) => {
+    setText(q);
+    getQuery(q);
+  };
+
+  return (
+    <div>
+      <form>
+        <input
+          type="text"
+          placeholder="search text"
+          value={text}
+          onChange={(e) => onChange(e.target.value)}
+          autoFocus
+        />
+      </form>
+    </div>
+  );
+};
+
+export default Search;
